modules/darkweb.py
import streamlit as st
import requests
import socks
import socket


def search_ahmia(query, proxy="127.0.0.1:9050"):
    """Busca en Ahmia vía Tor"""
    # Configurar proxy Tor
    socks.setdefaultproxy(socks.PROXY_TYPE_SOCKS5, "127.0.0.1", 9050)
    socket.socket = socks.socksocket

    url = f"http://juhanurmihxlp77nkq76byazcldy2hlmovfu2epvl5ankdibsot4csyd.onion/search/"
    params = {"q": query}
    # Implementa con requests y tor


def render_ui(user_id, api_keys, config, db):
    st.header("🧅 Dark Web / Onion Search")

    st.warning("⚠️ Requiere conexión Tor activa!")

    query = st.text_input("Buscar en dark web", placeholder="email@example.com")

    if st.button("Buscar en Ahmia"):
        if "Tor Proxy" in api_keys:
            # Implementa búsqueda
            st.info("Buscando...")
        else:
            st.error("Configura proxy Tor")